+++ lib/features/job/domain/entities/services_fee_system_entity.dart
@@ -4,7 +4,7 @@ class ServicesFeeSystemEntity {
   final int id;
   final String name;
   final String? description;
-  final int amount;
+  final double amount;
   final int? quantity;
 
   ServicesFeeSystemEntity({
@@ -21,7 +21,7 @@ class ServicesFeeSystemEntity {
       id: map['id'] ?? 0,
       name: map['name'] ?? '',
       description: map['description'], // Nullable field
-      amount: map['amount'] ?? 0,
+      amount: (map['amount'] ?? 0).toDouble(),
     );
   }
 
@@ -39,7 +39,7 @@ class ServicesFeeSystemEntity {
     int? id,
     String? name,
     String? description,
-    int? amount,
+    double? amount,
     int? quantity,
   }) {
     return ServicesFeeSystemEntity(
